+++ docs/app/pages/templates/[slug].vue
@@ -86,7 +86,7 @@ const images = computed(() => template.value
               color="neutral"
               trailing
               icon="i-ph-arrow-square-out"
-              to="template.demo"
+              :to="template.demo"
               target="_blank"
             />
 
@@ -97,7 +97,7 @@ const images = computed(() => template.value
                 target: '_blank',
               }, {
                 label: 'Clone on GitHub',
-                to: `https://github.com/${template.owner}/${template.name}`,
+                to: `https://github.com/${template.owner}/${template.name}/tree/${template.branch}`,
                 target: '_blank',
               }]"
             >