+++ docs/server/api/templates.json.get.ts
@@ -1,4 +1,18 @@
+import { handleCors } from 'h3'
+
 export default defineEventHandler(async (event) => {
+  const didHandleCors = handleCors(event, {
+    origin: ['https://nuxt.studio', 'https://dev.nuxt.studio'],
+    preflight: {
+      statusCode: 204,
+    },
+    methods: '*',
+  })
+
+  if (didHandleCors) {
+    return
+  }
+
   const templates = await queryCollection(event, 'templates').all()
 
   if (!templates) {