+++ test/bun.test.ts
@@ -1,6 +1,6 @@
 import { afterAll, describe, expect, test } from 'bun:test'
 import type { LocalDevelopmentDatabase } from '../src/types/index.ts'
-import { getLocalDatabase } from '../src/utils/sqlite.ts'
+import { getLocalDatabase } from '../src/utils/database.ts'
 
 describe('Local database', () => {
   let db: LocalDevelopmentDatabase
@@ -14,13 +14,13 @@ describe('Local database', () => {
   })
 
   test('load database', async () => {
-    db = await getLocalDatabase(':memory:')
+    db = await getLocalDatabase({ type: 'sqlite', filename: ':memory:' })
     expect(db).toBeDefined()
   })
 
   test('SQL: Create table', async () => {
     await db.exec('CREATE TABLE IF NOT EXISTS test (id TEXT PRIMARY KEY, name TEXT)')
-    const tables = await db.database?.all<{ name: string }>('SELECT name FROM sqlite_master WHERE type = ?', ['table'])
+    const tables = await db.database?.prepare('SELECT name FROM sqlite_master WHERE type = ?').all('table') as { name: string }[]
     expect(tables).toBeDefined()
     if (tables) {
       expect(tables.map(t => t.name)).toContain('test')
@@ -29,7 +29,7 @@ describe('Local database', () => {
 
   test('SQL: Insert data', async () => {
     await db.exec('INSERT INTO test (id, name) VALUES ("1", "Hello")')
-    const data = await db.database?.first<{ id: string, name: string }>('SELECT * FROM test')
+    const data = await db.database?.prepare('SELECT * FROM test').get() as { id: string, name: string }
     expect(data).toBeDefined()
     expect(data?.id).toBe('1')
     expect(data?.name).toBe('Hello')
@@ -37,13 +37,13 @@ describe('Local database', () => {
 
   test('SQL: Empty table', async () => {
     await db.exec('DELETE FROM test')
-    const data = await db.database?.first<{ id: string, name: string }>('SELECT * FROM test')
+    const data = await db.database?.prepare('SELECT * FROM test').get() as { id: string, name: string }
     expect(data).toBeNull()
   })
 
   test('SQL: Drop table', async () => {
     await db.exec('DROP TABLE test')
-    const tables = await db.database?.all<{ name: string }>('SELECT name FROM sqlite_master WHERE type = ?', ['table'])
+    const tables = await db.database?.prepare('SELECT name FROM sqlite_master WHERE type = ?').all('table') as { name: string }[]
     expect(tables).toBeDefined()
     if (tables) {
       expect(tables.map(t => t.name)).not.toContain('test')